07/11/2023_11:25_ana

+++ src/main/java/com/sisu/sisu/Dao/IPersonaDao.java
@@ -1,15 +1,13 @@
 package com.sisu.sisu.Dao;
 
-
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 
 import com.sisu.sisu.entitys.Persona;
 
-public interface IPersonaDao extends CrudRepository<Persona, Long> {
-    
-    
+public interface IPersonaDao extends CrudRepository<Persona, Integer> {
+
     @Query(value = "SELECT * FROM persona WHERE persona.ci = ?1", nativeQuery = true)
-public Persona findByCi(String ci);
+    public Persona findByCi(String ci);
 
 }
