+++ src/main/java/com/sisu/sisu/Dao/UsuarioDao.java
@@ -7,8 +7,8 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.sisu.sisu.entitys.Usuario;
 
-public interface UsuarioDao extends CrudRepository<Usuario, Long>{
+public interface UsuarioDao extends CrudRepository<Usuario, Integer> {
     @Query("SELECT u FROM Usuario u WHERE u.idUsuario = ?1")
-    Usuario buscarUsuarioPorId(Long id);
-    
+    Usuario buscarUsuarioPorId(Integer id);
+
 }
