+++ src/main/java/com/sisu/sisu/Service/UsuarioServiceImpl.java
@@ -10,39 +10,38 @@ import com.sisu.sisu.Dao.UsuarioDao;
 import com.sisu.sisu.entitys.Usuario;
 
 @Service
-public class UsuarioServiceImpl implements UsuarioService{
+public class UsuarioServiceImpl implements UsuarioService {
 
     @Autowired
     private UsuarioDao usuarioDao;
 
     @Override
     public List<Usuario> findAll() {
-       
+
         return (List<Usuario>) usuarioDao.findAll();
     }
 
     @Override
     public void save(Usuario usuario) {
-        
+
         usuarioDao.save(usuario);
     }
 
     @Override
-    public Usuario findOne(Long id) {
-        
+    public Usuario findOne(Integer id) {
+
         return usuarioDao.findById(id).orElse(null);
     }
 
     @Override
-    public void delete(Long id) {
-        
+    public void delete(Integer id) {
+
         usuarioDao.deleteById(id);
     }
 
     @Override
-    public Usuario buscarUsuarioPorId(Long id) {
+    public Usuario buscarUsuarioPorId(Integer id) {
         return usuarioDao.buscarUsuarioPorId(id);
     }
 
-    
 }
