+++ src/main/java/com/sisu/sisu/entitys/Persona.java
@@ -29,63 +29,63 @@ import lombok.Setter;
 @Entity
 @Table(name = "persona")
 public class Persona implements Serializable {
-	
-	private static final long serialVersionUID = 1L;
+
+    private static final long serialVersionUID = 1L;
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_persona")
-    private Long idPersona;
-	
+    private Integer idPersona;
+
     @Column(name = "nombres")
     private String nombres;
-    
+
     @Column(name = "ap_paterno")
     private String apPaterno;
-    
+
     @Column(name = "ap_materno")
     private String apMaterno;
-    
+
     @Column(name = "ci")
     private String ci;
 
-	@Column(name = "fecha_nacimiento")
-	@DateTimeFormat(pattern = "yyyy-MM-dd")
-	private LocalDate fecha_nac;
-    
+    @Column(name = "fecha_nacimiento")
+    @DateTimeFormat(pattern = "yyyy-MM-dd")
+    private LocalDate fecha_nac;
+
     @Column(name = "sexo")
     private String sexo;
-    
+
     @Column(name = "direccion")
     private String direccion;
-    
+
     @Column(name = "celular")
     private Integer celular;
-    
+
     @Column(name = "estado")
     private String estado;
-    
+
     @Column(name = "registro")
     @Temporal(TemporalType.TIMESTAMP)
     private Date registro;
-    
+
     @Column(name = "modificacion")
     @Temporal(TemporalType.TIMESTAMP)
     private Date modificacion;
-    
-    //--------------------------RELACION--------------------------------------
 
-	@OneToMany(cascade = CascadeType.ALL, mappedBy = "persona", fetch = FetchType.LAZY)
-	private List<Usuario> usuarios;
+    // --------------------------RELACION--------------------------------------
+
+    @OneToMany(cascade = CascadeType.ALL, mappedBy = "persona", fetch = FetchType.LAZY)
+    private List<Usuario> usuarios;
 
     @ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "idGradoAcademico")
+    @JoinColumn(name = "idGradoAcademico")
     private GradoAcademico grado_academico;
 
     @ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "idTipoEstadoCivil")
+    @JoinColumn(name = "idTipoEstadoCivil")
     private TiposEstadoCivil tipos_estado_civil;
 
     @ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "idDip")
+    @JoinColumn(name = "idDip")
     private Dip dip;
 }
